fixes to address pylint, @drewandersonnz and @portante reviews
