+++ ansible/playbooks/adhoc/grow_root_vg/filter_plugins
@@ -1,0 +1,1 @@
+../../../filter_plugins
