+++ ansible/playbooks/adhoc/grow_root_vg/grow_root_vg.yml
@@ -7,7 +7,6 @@
 #  To run:
 #  1. Source your AWS credentials (make sure it's the corresponding AWS account) into your environment
 #    export AWS_PROFILE=<aws account>
-#    export AWS_DEFAULT_REGION=<aws region>
 #
 # 2. run the playbook:
 #   ansible-playbook -e 'cli_tag_name=<tag-name>' grow_root_vg.yml
@@ -16,8 +15,8 @@
 #   ansible-playbook -e 'cli_tag_name=ops-compute-12345' grow_root_vg.yml
 #
 #  Notes:
-#  * By default this will do a900GB GP2 volume.  The can be overidden with the "-e 'cli_volume_size=something'" variable
-#  * This can be done with NO downtime on the host, but you probably ought to test a reboot 
+#  * By default this will do a 90GB GP2 volume.  The volume size can be overidden with the "-e 'cli_volume_size=something'" variable
+#  * This can be done with NO downtime on the host, but you probably ought to test a reboot. 
 #
 
 - name: Grow the rootvg volume group
@@ -29,15 +28,27 @@
   vars:
     cli_volume_size: 90
     cli_var_size: 64
+    os_size_min: 20
 
-  pre_tasks:
-  - fail:
+  tasks:
+  - name: "Check for required variables" 
+    fail:
       msg: "This playbook requires {{item}} to be set."
     when: "{{ item }} is not defined or {{ item }} == ''"
     with_items:
     - cli_tag_name
     - cli_volume_size
     - cli_var_size
+
+  - name: "Sanity check variables" 
+    fail: 
+      msg: "cli_volume_size should be at least cli_var_size + {{ os_size_min }}" 
+    when: "{{ cli_volume_size|int - cli_var_size|int }} < os_size_min|int "
+
+  - name: Need growpart script
+    yum:
+      name: cloud-utils-growpart
+      state: installed 
 
   - name: Attempt to find the Phyisical Volume that rootvg is using
     shell: "pvs | awk '/rootvg/ {print $1}'" 
@@ -64,13 +75,13 @@
 
   - name: set rootvg volume attributes
     set_fact:
-      rootvg: "{{ attached_volumes.volumes | vol_attrs(rootvg_pv.stdout) }}"
+      rootvg: "{{ rootvg_pv.stdout | vol_attrs }}"
  
   - debug: var=rootvg
 
   - name: get volume id of current rootvg volume
     set_fact:
-      rootvg_volume_id: "{{ attached_volumes.volumes | translate_volume_name(rootvg.device) }}"
+      rootvg_volume_id: "{{ attached_volumes.volumes | get_volume_id_by_linux_device(rootvg.device) }}"
 
   - debug: var=rootvg_volume_id
 
@@ -78,21 +89,39 @@
       msg: "Unable to look up the volume ID of {{ rootvg.device }}"
     when: rootvg_volume_id is undefined
 
+  - name: get volume size of current rootvg volume
+    set_fact:
+      rootvg_volume_size: "{{ attached_volumes.volumes | get_volume_size_by_linux_device(rootvg.device) }}"
+
+  - debug: var=rootvg_volume_size
+
   - name: find rootvg disk size before change 
     shell: "awk '/{{ rootvg.bare_device }}/ {if ($2==0) print $3}' /proc/partitions"
     register: orig_pv_size
 
   - debug: var=orig_pv_size
 
+  - name: Take snapshot
+    delegate_to: localhost
+    ec2_snapshot:
+      region: "{{ ec2_region }}"
+      volume_id: "{{ rootvg_volume_id }}"
+      description: "Snapshot of {{ cli_tag_name }} {{ rootvg.device }} before resize"
+      wait: no ## don't wait for the snapshot to complete pushing to S3, it still exists and waiting can take hours
+    when: cli_volume_size < rootvg_volume_size
+
   - name: Resize EBS volume 
     delegate_to: localhost
-    command: "aws ec2 modify-volume --volume-id {{ rootvg_volume_id }} --size {{ cli_volume_size }}"
+    ## afaict the ansible ec2 modules don't offer resize
+    command: "aws ec2 modify-volume --volume-id {{ rootvg_volume_id }} --size {{ cli_volume_size }} --region {{ ec2_region }}"
+    when: cli_volume_size < rootvg_volume_size
     register: resize_volume
+
+  - debug: var=resize_volume
 
   - wait_for: 
       timeout: 10 
-
-  - debug: var=resize_volume
+    when: cli_volume_size < rootvg_volume_size
 
   - name: Fail when problems extending
     fail:
@@ -109,14 +138,9 @@
   - debug: var=new_pv_size
 
   - name: check if rootvg disk size changed
-    fail:
+    debug:
       msg: Rootvg disk size was not changed
     when: orig_pv_size.stdout == new_pv_size.stdout
-
-  - name: Need growpart script
-    yum:
-      name: cloud-utils-growpart
-      state: installed 
 
   - name: Run growpart on last partition 
     shell: "growpart {{ rootvg.device }} $(grep -c {{ rootvg.bare_device }}[0-9] /proc/partitions) -v"
